requirements.txt
Flask==2.1.1
numpy==1.22.3
opencv_python==4.5.5.64
python_fcl==0.6.1
PyYAML==6.0
requests==2.27.1
